+++ example/main.tf
@@ -8,6 +8,7 @@ module "postgres" {
   # nomad
   nomad_datacenters = ["dc1"]
   nomad_namespace   = "default"
+  nomad_host_volume = "persistence"
 
   # postgres
   service_name                    = "postgres"
@@ -16,5 +17,7 @@ module "postgres" {
   admin_user                      = data.vault_generic_secret.postgres_secrets.data.username
   admin_password                  = data.vault_generic_secret.postgres_secrets.data.password
   database                        = "metastore"
-  container_environment_variables = ["PGDATA=/var/lib/postgresql/data"]
+  volume_destination              = "/var/lib/postgresql/data"
+  use_host_volume                 = true
+  container_environment_variables = ["PGDATA=/var/lib/postgresql/data/"]
 }