+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 ### Changed
 - Renamed vars in `variables.tf` and where they are used #8
-- Added host volume for postgres #12
 
 ### Added
 - Using Vault to generate username and password for postgres with kv #13
+- Added host volume for postgres and dynamic variables #12
+- Added switch for host volume #12
 
 ## [0.0.1]